medium

+++ src/Solution.java
@@ -66,7 +66,7 @@ class Solver{
     }
     boolean eq(Integer[] a, Integer[] b){
         for (int i = 0; i < a.length; i++) {
-            if(a[i] != b[i]){
+            if((int)a[i] != b[i]){
                 return false;
             }
         }
